Update page's CSS and router
 - 페이지의 CSS , grid system 변경
 - TODO : router의 path에 따른 router 분리를 한다.
 - TODO : firestore을 사용하는 router을 분리한다.
