+++ components/footer.ejs
@@ -1,11 +1,6 @@
-<div class="row">
+<div class="fixed-bottom">
+  <div class="text-center">
+    2020 Duckey.Kim :<a href="mailto:">duckey.kim@gmail.com</a>
+  </div>
 
-  <div class="col-12 col-md-12">
-    Copyright(c) 2020 Duckey.Kim
-  </div>
-  <div class="col-12 col-md-12"><a href="mailto:#">duckey-kdh@google.com</a>
-  </div>
-  <div class="col-12 col-md-12"><a href="https://github.com/duckey-kim"
-      style="text-decoration: none; color: brown;">https://github.com/duckey-kim</a>
-  </div>
 </div>
