+++ views/boards/index.ejs
@@ -1,20 +1,25 @@
 <div class="row">
   <% data.forEach((posts,category)=>{%>
   <div class="col-12 col-md-6">
-    <div class="col-12 col-md-12 borded rounded bg-bluesky" id="indexCategory"
-      onclick="location.href='/boards/<%= category %>'">
-      <%= category.toUpperCase()%>
-    </div>
-    <% posts.forEach((post)=>{ %>
-    <div class="row" id="boards" onclick="location.href='/boards/<%=category%>/posts/<%= post.postnum%>'">
-      <div class="col-8 col-md-8 text-truncate">
-        <%=post.posttitle%>
+    <div class="indexBox mt-1">
+      <div class="row borded rounded bg-pink indexCategory" onclick="location.href='/boards/<%= category %>/posts'">
+        <%= category.toUpperCase()%>
       </div>
-      <div class="col-4 col-md-4 text-right text-truncate">
-        <%=post.postuser%>
+      <div class="row">
+        <div class="col-12">
+          <% posts.forEach((post)=>{ %>
+          <div class="row mt-1" id="boards" onclick="location.href='/boards/<%=category%>/posts/<%= post.postnum%>'">
+            <div class="col-8 text-truncate">
+              <%=post.posttitle%>
+            </div>
+            <div class="col-4 text-right text-truncate">
+              <%=post.postuser%>
+            </div>
+          </div>
+          <% }) %>
+        </div>
       </div>
     </div>
-    <% }) %>
   </div>
   <% })%>
 </div>
